requirements.txt
discord.py
expiring-dict
aiosqlite